Add FSI driver
FSI is the Flexible Service Interface, an i2c-like bus used on IBM POWER
processors for boot control, monitoring and debug. The ASPEED AST2600
contains a FSI master that can be used to talk to slaves on this bus.

The u-boot driver is a light porting of the work in progress Linux
driver. It allows basic reading and writing across the FSI link, as well
as control of the bus speed divisor.

Signed-off-by: Joel Stanley <joel@jms.id.au>
---
v2: remove stray printf
  fix status output
  fix trace satements
  fix setting of divisor