aspeed: clock: Add FSI clock
The clock is used by both the FSI masters in the AST2600. It must not be
reset if it is already enabled, such as when probing the device driver
for the second master, as that will lose state for the first.

Signed-off-by: Joel Stanley <joel@jms.id.au>
---
v2: don't bounce when already enabled

don't bounce clock

Signed-off-by: Joel Stanley <joel@jms.id.au>