configs/ast2600: Malloc pool can be all of sram
The size of the early (pre-DRAM) SRAM heap in u-boot proper is extended
to 88KB, all of the regular SRAM region on A1 and up.

Testing found that the DRAM driver would perform an allocation that
exceeded the limits. As SRAM is unused when u-boot proper is running,
it can all be allocated.

As MMC boot is not functional on the A0, only systems with the extended
SRAM size will be running the MMC configuration.

The SPL early heap is fixed to 0x800, the default before this change.
Testing shows that:

  SPL malloc() before relocation used 0x794 bytes (1 KB)

So 2KB should be enough.

Signed-off-by: Joel Stanley <joel@jms.id.au>