powerpc: Add basic support for the Microwatt CPU
Microwatt is a 64-bit Power ISA soft core.  This adds support for
Microwatt running on an Arty A7-100 FPGA development board.  For
now, U-Boot runs on it in 64-bit little-endian mode with the MMU
turned off.  U-Boot is stored in SPI flash and is loaded into DRAM
by the existing firmware, which does DRAM initialization and then
copies an ELF image from SPI flash into DRAM and jumps to it.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>